// src/types/childrenForm.ts
export interface IChildrensForm {
  firstName: string;
  lastName: string;
  dob: string;
  gender: string;
  guardianName: string;
  email: string;
  phoneNumber: string;
  country: string;
  schoolName: string;
  grade: string;
  promotionals: boolean;
  concerns: string;
  HowLongHasYourChildBeenExperiencingTheseChallenges: string;
  WhySeekTherapy: string;
  ChangesInBehaviour: string;
  LivingSituation: string;
  FamilyDynamics: string;
  SupportNetwork: string;
  SchoolReport: string;
  AcademicPerformance: string;
  MentalHealthSupport: string;
  Medications: string;
  Outcome: string;
  BehaviorFocus: string;
  LanguagePreference: string;
  referralCode: string;
}